backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
openai>=1.12.0
httpx>=0.27.0
faiss-cpu>=1.7.4
pandas>=2.0.0
numpy>=1.26.0
python-dotenv==1.0.0
python-multipart==0.0.6
